% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatPost.R
\name{FormatPost}
\alias{FormatPost}
\title{Format an R Markdown post for a Jekyll blog}
\usage{
FormatPost(rmd, knitr.files.dir = "knitr_files", tmp.dir = "_built")
}
\arguments{
\item{rmd}{The name (path) to specific Rmd file to convert.}

\item{knitr.files.dir}{The permanent directory where
the files like plots are placed.}

\item{tmp.dir}{Temporary directory for the temporary html file.}
}
\value{
The name (path) of the new Markdown post
}
\description{
Format an R Markdown post for a Jekyll blog
and place the resulting Markdown post in directory '_posts'.
}

