% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatPost.R
\name{FormatPost}
\alias{FormatPost}
\title{Format an R Markdown post for a Jekyll blog}
\usage{
FormatPost(rmd, knitr.files.dir = "knitr_files",
  date.format = "\%B \%d, \%Y", date.locale = readr::locale(tz =
  "US/Central"))
}
\arguments{
\item{rmd}{The name (path) to specific Rmd file to convert.}

\item{knitr.files.dir}{The permanent directory where
the files like plots are placed.}

\item{date.format}{Format of the date used in the YAML header.
Default corresponds for example to "August 15, 2016".}

\item{date.locale}{A locale object, defining the defaults of a country.}
}
\value{
The name (path) of the new Markdown post.
}
\description{
Format an R Markdown Pretty Document (package prettydoc) post for a Jekyll
blog and place the resulting Markdown post in directory '_posts'.
}
\seealso{
readr::parse_date
}
